# shared/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Data Voyager API
  version: 0.1.0
  description: Data Voyager REST API
servers:
  - url: /api/v1
    description: API v1
...
paths:
  /connections:
    get:
      operationId: listConnections
      summary: List all connections
      tags: [connections]
      parameters:
        - in: query
          name: type
          schema:
            type: string
          description: Filter by datasource type
        - in: query
          name: is_active
          schema:
            type: boolean
          description: Filter by active status
        - in: query
          name: created_by
          schema:
            type: string
          description: Filter by creator
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionListResponse"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createConnection
      summary: Create a connection
      tags: [connections]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /connections/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: getConnection
      summary: Get a connection by ID
      tags: [connections]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      operationId: updateConnection
      summary: Update a connection
      tags: [connections]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConnectionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteConnection
      summary: Delete a connection
      tags: [connections]
      responses:
        "204":
          description: No Content
        "500":
          $ref: "#/components/responses/InternalError"
  /connections/test:
    post:
      operationId: testConnectionConfig
      summary: Test a connection config without saving
      tags: [connections]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestConnectionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /connections/{id}/test:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: testConnection
      summary: Test a connection
      tags: [connections]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /connections/{id}/schema:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getConnectionSchema
      summary: Get datasource schema for a connection
      tags: [connections]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionSchemaResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /connections/{id}/query:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: queryConnection
      summary: Execute a query through a connection
      tags: [connections]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "501":
          $ref: "#/components/responses/NotImplemented"
        "502":
          $ref: "#/components/responses/BadGateway"
  /connections/{id}/query/batch:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: batchQueryConnection
      summary: Execute multiple queries through a connection and return all results
      tags: [connections]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchQueryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchQueryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "501":
          $ref: "#/components/responses/NotImplemented"
        "502":
          $ref: "#/components/responses/BadGateway"
  /connection-types:
    get:
      operationId: listConnectionTypes
      summary: List supported connection types
      tags: [connections]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTypesResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /connection-stats:
    get:
      operationId: getConnectionStats
      summary: Get connection statistics
      tags: [connections]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionStatsResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
    Connection:
      type: object
      required: [id, name, type, config, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          description: Datasource type identifier (e.g. "postgresql", "clickhouse")
        config:
          type: object
          additionalProperties: true
          description: Driver-specific connection configuration
          x-go-type: json.RawMessage
        description:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
        tags:
          type: array
          items:
            type: string
    ConnectionTestResult:
      type: object
      required: [is_connected, message]
      properties:
        is_connected:
          type: boolean
        message:
          type: string
        latency_ms:
          type: integer
          format: int64
        tested_at:
          type: string
          format: date-time
    ConnectionStats:
      type: object
      required: [total_count, active_count, count_by_type]
      properties:
        total_count:
          type: integer
          format: int64
        active_count:
          type: integer
          format: int64
        count_by_type:
          type: object
          additionalProperties:
            type: integer
            format: int64
    CreateConnectionRequest:
      type: object
      required: [name, type, config]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          minLength: 1
        config:
          type: object
          additionalProperties: true
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        created_by:
          type: string
    UpdateConnectionRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        config:
          type: object
          additionalProperties: true
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        is_active:
          type: boolean
    TestConnectionRequest:
      type: object
      required: [type, config]
      properties:
        type:
          type: string
          minLength: 1
        config:
          type: object
          additionalProperties: true
    TimeRange:
      type: object
      properties:
        from:
          type: string
          description: >
            Start of the time range. Accepts ISO 8601 ("2024-04-01T00:00:00Z"),
            Unix ms as a numeric string ("1712160000000"),
            or a relative expression ("1 Hours ago").
        to:
          type: string
          description: >
            End of the time range. Accepts the same formats as "from", plus "Now".
    QueryRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          description: >
            Raw query template. Server-side {{ variable }} substitution is
            applied before execution. Built-in variables (__ prefix) are
            injected automatically from time_range.
        variables:
          type: object
          additionalProperties: true
          description: User-defined variables for {{ }} template substitution.
        time_range:
          $ref: "#/components/schemas/TimeRange"
        limit:
          type: integer
          default: 10000
    QueryStats:
      type: object
      required: [execution_time_ms, rows_returned]
      properties:
        execution_time_ms:
          type: integer
          format: int64
        rows_returned:
          type: integer
          format: int64
        bytes_read:
          type: integer
          format: int64
    QueryInspect:
      type: object
      required: [raw_query, executed_query]
      properties:
        raw_query:
          type: string
          description: Original template before variable substitution.
        executed_query:
          type: string
          description: Final query after all variable substitution.
        variables:
          type: object
          additionalProperties: true
          description: Full context used for substitution (including built-in __ variables).
    FieldKind:
      type: string
      description: Semantic type of a field, used for rendering (axis selection, formatting).
      enum: [time, number, string, boolean]
    FrameType:
      type: string
      description: Hint for how the DataFrame should be visualized.
      enum: [table, time_series, logs]
    Field:
      type: object
      required: [name, kind, values]
      properties:
        name:
          type: string
        kind:
          $ref: "#/components/schemas/FieldKind"
        type:
          type: string
          description: Native database type string (display only).
        labels:
          type: object
          additionalProperties:
            type: string
          description: Key-value labels attached to the field (e.g. Prometheus metric labels).
        values:
          type: array
          items: {}
    DataFrame:
      type: object
      required: [frame_type, fields]
      properties:
        name:
          type: string
        frame_type:
          $ref: "#/components/schemas/FrameType"
        fields:
          type: array
          items:
            $ref: "#/components/schemas/Field"
    QueryResult:
      type: object
      required: [frames]
      properties:
        frames:
          type: array
          items:
            $ref: "#/components/schemas/DataFrame"

    # Response envelopes
    ConnectionResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Connection"
    ConnectionListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Connection"
    ConnectionTestResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ConnectionTestResult"
    ConnectionTypesResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            type: string
    ConnectionStatsResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ConnectionStats"
    ConnectionSchemaResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: true
    QueryResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/QueryResult"
        stats:
          $ref: "#/components/schemas/QueryStats"
        inspect:
          $ref: "#/components/schemas/QueryInspect"
    BatchQueryItem:
      type: object
      required: [ref_id, request]
      properties:
        ref_id:
          type: string
          description: Query identifier (e.g. "A", "B"). Returned in results.
        request:
          $ref: "#/components/schemas/QueryRequest"
    BatchQueryRequest:
      type: object
      required: [queries]
      properties:
        queries:
          type: array
          items:
            $ref: "#/components/schemas/BatchQueryItem"
          minItems: 1
    BatchQueryResultItem:
      type: object
      required: [ref_id]
      properties:
        ref_id:
          type: string
        data:
          $ref: "#/components/schemas/QueryResult"
        stats:
          $ref: "#/components/schemas/QueryStats"
        inspect:
          $ref: "#/components/schemas/QueryInspect"
        error:
          type: string
          description: Per-query error message if this query failed.
    BatchQueryResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/BatchQueryResultItem"

    # AI settings schemas
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotImplemented:
      description: Not Implemented
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: Bad Gateway — upstream datasource connection or query failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"